Refactor - move model utils file to utils folder
