+++ back-end/models/users.model.js
@@ -1,5 +1,5 @@
 const { HttpErrors } = require("../../shared/HttpErrors");
-const { rejectWhenNonExistent, gQuerier } = require("./model-utils.js");
+const { rejectWhenNonExistent, gQuerier } = require("./utils/model-utils.js");
 const { updateUserTopics } = require("./user-topic-join.model");
 const { selectUserTopics } = require("./user-topic-join.model");
 const { selectTopicAndInsertIfNonExistent } = require("./topicss.model");
